refactor: api change. [board,chat,find,login,logout,oauth,pay,signup]
