--- src/app/register/_lib/getTest.ts
@@ -1,7 +1,0 @@
-export const getTest = async () => {
-  const data = await fetch(`/api/test`, {
-    method: 'GET'
-  });
-
-  return data.json();
-};
